+++ packages/coast_audio/lib/src/node/abstract/data_source_node.dart
@@ -1,6 +1,10 @@
 import 'package:coast_audio/coast_audio.dart';
 
+/// [DataSourceNode] is a node that provides audio data to outputBus.
 abstract class DataSourceNode extends AudioNode with SingleOutNodeMixin {
+  /// The format of the audio data that this node provides.
+  ///
+  /// If this node provides audio data with multiple formats, you should return null.
   AudioFormat? get outputFormat;
 
   @override