+++ packages/coast_audio/lib/src/node/bus/audio_bus_connection_exception.dart
@@ -1,0 +1,19 @@
+class AudioBusConnectionException implements Exception {
+  const AudioBusConnectionException(this.message, this.code);
+
+  const AudioBusConnectionException.sameNode()
+      : message = 'input bus and output bus have same node reference',
+        code = -1;
+
+  const AudioBusConnectionException.incompatibleFormat()
+      : message = 'input bus and output bus have incompatible format',
+        code = -2;
+
+  final String message;
+  final int code;
+
+  @override
+  String toString() {
+    return message;
+  }
+}