+++ packages/coast_audio/lib/src/node/bus/audio_output_bus.dart
@@ -1,7 +1,9 @@
 import 'package:coast_audio/coast_audio.dart';
+import 'package:coast_audio/src/node/bus/audio_input_bus.dart';
 
 typedef OutputFormatResolver = AudioFormat? Function(AudioOutputBus bus);
 
+/// [AudioOutputBus] represents a audio node's output format and connection.
 class AudioOutputBus extends AudioBus {
   AudioOutputBus({
     required AudioNode node,
@@ -15,16 +17,46 @@ class AudioOutputBus extends AudioBus {
   AudioFormat? resolveFormat() => _formatResolver(this);
 
   AudioInputBus? _connectedBus;
+
+  /// The connected [AudioInputBus] of this bus.
+  ///
+  /// [connectedBus]'s format must be the same as this bus's format.
   AudioInputBus? get connectedBus => _connectedBus;
 
-  void onConnect(AudioInputBus bus) {
-    _connectedBus = bus;
+  void connect(AudioInputBus inputBus) {
+    if (node == inputBus.node) {
+      throw const AudioBusConnectionException.sameNode();
+    }
+
+    if (!canConnect(inputBus)) {
+      throw const AudioBusConnectionException.incompatibleFormat();
+    }
+
+    _connectedBus = inputBus;
+    inputBus.onConnect(this);
   }
 
-  void onDisconnect() {
+  void disconnect() {
+    _connectedBus?.onDisconnect();
     _connectedBus = null;
   }
 
+  bool canConnect(AudioInputBus inputBus) {
+    if (node == inputBus.node) {
+      return false;
+    }
+
+    final outputFormat = resolveFormat();
+    final inputFormat = inputBus.resolveFormat();
+
+    if (outputFormat == null || inputFormat == null) {
+      return true;
+    }
+
+    return outputFormat.isSameFormat(inputFormat);
+  }
+
+  /// Read audio data from the associated node.
   AudioReadResult read(AudioBuffer buffer) {
     return node.read(this, buffer);
   }