+++ packages/coast_audio/lib/src/node/decoder_node.dart
@@ -10,15 +10,9 @@ class DecoderNode extends DataSourceNode {
   @override
   AudioFormat get outputFormat => decoder.outputFormat;
 
-  final _listeners = <DecodeResultListener>[];
-
   @override
   AudioReadResult read(AudioOutputBus outputBus, AudioBuffer buffer) {
     final result = decoder.decode(destination: buffer);
-    for (var listener in _listeners) {
-      listener(result);
-    }
-
     return AudioReadResult(frameCount: result.frames, isEnd: result.isEnd);
   }
 }