+++ packages/coast_audio/lib/src/node/mixin/disposable_node_mixin.dart
@@ -1,9 +1,11 @@
 import 'package:coast_audio/coast_audio.dart';
 import 'package:meta/meta.dart';
 
+/// A mixin that provides a [dispose] method to dispose the node's resources.
 mixin SyncDisposableNodeMixin on AudioNode implements SyncDisposable {
   var _isDispose = false;
 
+  /// Returns true if the node is disposed.
   @override
   bool get isDisposed => _isDispose;
 
@@ -13,6 +15,10 @@ mixin SyncDisposableNodeMixin on AudioNode implements SyncDisposable {
       throw DisposedException(this, target);
     }
   }
+
+  /// Dispose the node's resources.
+  ///
+  /// This method should be called when the node is no longer needed.
 
   @override
   @mustCallSuper