+++ packages/coast_audio/lib/src/node/mixin/processor_node_mixin.dart
@@ -1,5 +1,6 @@
 import 'package:coast_audio/coast_audio.dart';
 
+/// An audio node that reads from an input bus and calls [process] method automatically.
 mixin ProcessorNodeMixin on SingleInNodeMixin {
   @override
   AudioReadResult read(AudioOutputBus outputBus, AudioBuffer buffer) {
@@ -10,5 +11,9 @@ mixin ProcessorNodeMixin on SingleInNodeMixin {
     );
   }
 
+  /// Process the input buffer and return the result.
+  ///
+  /// The [buffer] is the input buffer to process.
+  /// The [isEnd] is true if the input buffer is the last buffer of the input stream.
   AudioReadResult process(AudioBuffer buffer, bool isEnd);
 }